extracted ai schema stubs into external directory, removed unused enum
